refactor book and label class in accordance to solid principles
