+++ labels.json
@@ -1,12 +1,12 @@
 [
   {
-    "id": 409,
-    "title": "secrets of life",
-    "color": "red"
+    "id": 270,
+    "title": "Harry Potter",
+    "color": "black"
   },
   {
-    "id": 612,
-    "title": "the oasis",
+    "id": 285,
+    "title": "Hitchhiker's Guide",
     "color": "blue"
   }
 ]
